aimkg-recommender-UI/requirements.txt
Flask
neo4j
python-dotenv
pandas
sentence-transformers
torch